+++ domain/src/main/java/org/oppia/domain/topic/TopicController.kt
@@ -2,6 +2,7 @@ package org.oppia.domain.topic
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
+import org.json.JSONArray
 import org.oppia.app.model.ChapterPlayState
 import java.lang.IllegalArgumentException
 import javax.inject.Inject
@@ -19,7 +20,8 @@ import org.oppia.app.model.Translation
 import org.oppia.app.model.TranslationMapping
 import org.oppia.app.model.Voiceover
 import org.oppia.app.model.VoiceoverMapping
-import org.oppia.domain.exploration.ExplorationRetriever
+import org.oppia.domain.util.JsonAssetRetriever
+import org.oppia.domain.util.StateRetriever
 import org.oppia.util.data.AsyncResult
 import org.oppia.util.data.DataProvider
 import org.oppia.util.data.DataProviders
@@ -32,6 +34,9 @@ const val TEST_SKILL_CONTENT_ID_1 = "test_skill_content_id_1"
 const val TEST_QUESTION_ID_0 = "question_id_0"
 const val TEST_QUESTION_ID_1 = "question_id_1"
 const val TEST_QUESTION_ID_2 = "question_id_2"
+const val TEST_QUESTION_ID_3 = "question_id_3"
+const val TEST_QUESTION_ID_4 = "question_id_4"
+const val TEST_QUESTION_ID_5 = "question_id_5"
 
 private const val QUESTION_DATA_PROVIDER_ID = "QuestionDataProvider"
 
@@ -39,7 +44,8 @@ private const val QUESTION_DATA_PROVIDER_ID = "QuestionDataProvider"
 @Singleton
 class TopicController @Inject constructor(
   private val dataProviders: DataProviders,
-  private val explorationRetriever: ExplorationRetriever
+  private val jsonAssetRetriever: JsonAssetRetriever,
+  private val stateRetriever: StateRetriever
 ) {
   /** Returns the [Topic] corresponding to the specified topic ID, or a failed result if no such topic exists. */
   fun getTopic(topicId: String): LiveData<AsyncResult<Topic>> {
@@ -79,49 +85,33 @@ class TopicController @Inject constructor(
   }
 
   fun retrieveQuestionsForSkillIds(skillIdsList: List<String>): DataProvider<List<Question>> {
-    return dataProviders.createInMemoryDataProviderAsync(QUESTION_DATA_PROVIDER_ID) {
-      loadQuestionsForSkillIds(skillIdsList)
-    }
+     return dataProviders.createInMemoryDataProvider(QUESTION_DATA_PROVIDER_ID) {
+        loadQuestionsForSkillIds(skillIdsList)
+      }
   }
 
   // Loads and returns the questions given a list of skill ids.
-  @Suppress("RedundantSuspendModifier") // DataProviders expects this function to be a suspend function.
-  private suspend fun loadQuestionsForSkillIds(skillIdsList: List<String>): AsyncResult<List<Question>> {
-    return try {
-      AsyncResult.success(loadQuestions(skillIdsList))
-    } catch (e: Exception) {
-      AsyncResult.failed(e)
-    }
+  private fun loadQuestionsForSkillIds(skillIdsList: List<String>): List<Question> {
+    return loadQuestions(skillIdsList)
   }
 
-  @Suppress("RedundantSuspendModifier") // Force callers to call this on a background thread.
-  private suspend fun loadQuestions(skillIdsList: List<String>): List<Question> {
+  private fun loadQuestions(skillIdsList: List<String>): List<Question> {
     val questionsList = mutableListOf<Question>()
-    val questionsJSON = explorationRetriever.loadJsonFromAsset(
+    val questionsJSON = jsonAssetRetriever.loadJsonFromAsset(
       "sample_questions.json")?.getJSONArray("questions")
     for (skillId in skillIdsList) {
       when (skillId) {
-        TEST_SKILL_ID_0 -> questionsList.add(
-          Question.newBuilder()
-            .setQuestionId(TEST_QUESTION_ID_0)
-            .setQuestionState(explorationRetriever.createStateFromJson(
-              "question",questionsJSON?.getJSONObject(0)))
-            .addLinkedSkillIds(TEST_SKILL_ID_0)
-            .build())
-        TEST_SKILL_ID_1 -> questionsList.add(
-          Question.newBuilder()
-            .setQuestionId(TEST_QUESTION_ID_1)
-            .setQuestionState(explorationRetriever.createStateFromJson(
-              "question",questionsJSON?.getJSONObject(1)))
-            .addLinkedSkillIds(TEST_SKILL_ID_1)
-            .build())
-        TEST_SKILL_ID_2 -> questionsList.add(
-          Question.newBuilder()
-            .setQuestionId(TEST_QUESTION_ID_2)
-            .setQuestionState(explorationRetriever.createStateFromJson(
-              "question",questionsJSON?.getJSONObject(2)))
-            .addLinkedSkillIds(TEST_SKILL_ID_2)
-            .build())
+        TEST_SKILL_ID_0 -> questionsList.addAll(mutableListOf(
+          createTestQuestion0(questionsJSON),
+          createTestQuestion1(questionsJSON),
+          createTestQuestion2(questionsJSON)))
+        TEST_SKILL_ID_1 -> questionsList.addAll(mutableListOf(
+          createTestQuestion0(questionsJSON),
+          createTestQuestion3(questionsJSON)))
+        TEST_SKILL_ID_2 -> questionsList.addAll(mutableListOf(
+        createTestQuestion2(questionsJSON),
+        createTestQuestion4(questionsJSON),
+        createTestQuestion5(questionsJSON)))
         else -> {
           throw IllegalStateException("Invalid skill ID: $skillId")
         }
@@ -129,6 +119,62 @@ class TopicController @Inject constructor(
     }
     return questionsList
   }
+
+  private fun createTestQuestion0(questionsJSON: JSONArray?): Question {
+    return Question.newBuilder()
+      .setQuestionId(TEST_QUESTION_ID_0)
+      .setQuestionState(stateRetriever.createStateFromJson(
+        "question",questionsJSON?.getJSONObject(0)))
+       .addAllLinkedSkillIds(mutableListOf(TEST_SKILL_ID_0, TEST_SKILL_ID_1))
+      .build()
+  }
+
+  private fun createTestQuestion1(questionsJSON: JSONArray?): Question {
+    return Question.newBuilder()
+      .setQuestionId(TEST_QUESTION_ID_1)
+      .setQuestionState(stateRetriever.createStateFromJson(
+        "question",questionsJSON?.getJSONObject(1)))
+      .addAllLinkedSkillIds(mutableListOf(TEST_SKILL_ID_0))
+      .build()
+  }
+
+  private fun createTestQuestion2(questionsJSON: JSONArray?): Question {
+    return Question.newBuilder()
+      .setQuestionId(TEST_QUESTION_ID_2)
+      .setQuestionState(stateRetriever.createStateFromJson(
+        "question",questionsJSON?.getJSONObject(2)))
+      .addAllLinkedSkillIds(mutableListOf(TEST_SKILL_ID_0, TEST_SKILL_ID_2))
+      .build()
+  }
+
+  private fun createTestQuestion3(questionsJSON: JSONArray?): Question {
+    return Question.newBuilder()
+      .setQuestionId(TEST_QUESTION_ID_3)
+      .setQuestionState(stateRetriever.createStateFromJson(
+        "question",questionsJSON?.getJSONObject(0)))
+      .addAllLinkedSkillIds(mutableListOf(TEST_SKILL_ID_1))
+      .build()
+  }
+
+  private fun createTestQuestion4(questionsJSON: JSONArray?): Question {
+    return Question.newBuilder()
+      .setQuestionId(TEST_QUESTION_ID_4)
+      .setQuestionState(stateRetriever.createStateFromJson(
+        "question",questionsJSON?.getJSONObject(1)))
+      .addAllLinkedSkillIds(mutableListOf(TEST_SKILL_ID_2))
+      .build()
+  }
+
+  private fun createTestQuestion5(questionsJSON: JSONArray?): Question {
+    return Question.newBuilder()
+      .setQuestionId(TEST_QUESTION_ID_5)
+      .setQuestionState(stateRetriever.createStateFromJson(
+        "question",questionsJSON?.getJSONObject(2)))
+      .addAllLinkedSkillIds(mutableListOf(TEST_SKILL_ID_2))
+      .build()
+  }
+
+
 
   private fun createTestTopic0(): Topic {
     return Topic.newBuilder()
